move restic messages to a restic package
